requirements.txt
flask==3.0.2
flask-sqlalchemy==3.1.1
flask-login==0.6.3
flask-wtf==1.2.1
google-generativeai==0.3.2
python-dotenv==1.0.1
werkzeug==3.0.1
email-validator==2.1.0.post1 